pyproject.toml
[project]
name = "test-adk"
version = "0.1.0"
description = "VNStock Agent với Google ADK - Tích hợp 32 MCP tools để truy vấn dữ liệu chứng khoán Việt Nam"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.18.0",
    "httpx>=0.24.0",  # HTTP client để gọi MCP server
    "pyyaml>=6.0",  # Đọc config YAML
    "fastapi>=0.104.0",  # FastAPI web framework
    "uvicorn[standard]>=0.24.0",  # ASGI server
    "pydantic>=2.0.0",  # Data validation
    "pydantic-settings>=2.0.0",  # Settings management
    "python-dotenv>=1.0.1",  # Load .env cho GOOGLE_API_KEY
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
